apollo_infra: organize tests
